Rename file & add back return characters
