+++ src/components/EndUserLicenseAgreement/index.jsx
